Provide the menu as a module.
Split the menu functionality of main.py out into a module pew_menu that can be used by applications.
It exports these functions:
- menu: Synchronously runs the menu.
- menugen: Generator that allows background processing while the menu is running, in particular the list of menu entries may be modified live.
